separation of concerns with components folder
